fix(core): extract CLI error classification to satisfy file length gate
Move _resolve_cli_worker_error and scrub helpers from cli_streaming_parser
into cli_error_classify.py (QG max 300 lines).
